gateway/requirements.txt
fastapi
uvicorn
httpx
pydantic
PyYAML
python-dotenv
pyyaml
pydantic